Use tuples instead of CompositeReporter
