+++ ndc-test/src/lib.rs
@@ -21,7 +21,7 @@ use ndc_client::models::{self};
 use error::Result;
 
 use rand::SeedableRng;
-use reporter::{Reporter, TestResults};
+use reporter::Reporter;
 use serde::de::DeserializeOwned;
 use snapshot::{snapshot_test, SnapshottingConnector};
 
@@ -42,26 +42,6 @@ impl Connector for Configuration {
     async fn mutation(&self, request: models::MutationRequest) -> Result<models::MutationResponse> {
         Ok(api::mutation_post(self, request).await?)
     }
-}
-
-pub fn report(results: &TestResults) -> String {
-    use colored::Colorize;
-
-    let mut result = format!("Failed with {0} test failures:", results.failures.len())
-        .red()
-        .to_string();
-
-    let mut ix = 1;
-    for failure in results.failures.iter() {
-        result += format!("\n\n[{0}] {1}", ix, failure.name).as_str();
-        for path_element in failure.path.iter() {
-            result += format!("\n  in {0}", path_element).as_str();
-        }
-        result += format!("\nDetails: {0}", failure.error).as_str();
-        ix += 1;
-    }
-
-    result
 }
 
 pub async fn test_connector<C: Connector, R: Reporter>(
